neplweb01/requirements.txt
asgiref==3.4.1
Django==3.2.9
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
paho-mqtt==1.6.1
